+++ scripts/provision/debian/custom/puppet_install_dist/A_base.sh
@@ -7,15 +7,17 @@ set -x
 setup_hiera_eyaml=$ENV_puppet_install_dist_setup_hiera_eyaml
 [[ -z "$setup_hiera_eyaml" ]] && setup_hiera_eyaml=false
 
+puppet_packages=$ENV_puppet_install_dist_puppet_packages
+[[ -z "$puppet_packages" ]] && puppet_packages='puppet ruby-deep-merge'
+
 if [[ -z "$(which puppet)" ]]; then
   wget -O /tmp/puppetlabs-release.deb https://apt.puppetlabs.com/puppetlabs-release-$(lsb_release -cs).deb
   dpkg -i /tmp/puppetlabs-release.deb
 
   apt-get update
-  DEBIAN_FRONTEND=noninteractive apt-get -yV install puppet ruby-deep-merge
-else
-  DEBIAN_FRONTEND=noninteractive apt-get -yV install ruby-deep-merge
 fi
+
+DEBIAN_FRONTEND=noninteractive apt-get -yV install $puppet_packages
 
 if [[ "$setup_hiera_eyaml" == 'true' ]]; then
   gem install --verbose hiera-eyaml